@@ -28,12 +28,12 @@
         "source": "https://github.com/wintercms/winter"
     },
     "require": {
-        "php": ">=7.2.9",
-        "winter/storm": "~1.1.2",
-        "winter/wn-system-module": "~1.1.2",
-        "winter/wn-backend-module": "~1.1.2",
-        "winter/wn-cms-module": "~1.1.2",
-        "laravel/framework": "~6.0",
+        "php": "^8.0.2",
+        "winter/storm": "dev-wip/1.2 as 1.2",
+        "winter/wn-system-module": "dev-wip/1.2",
+        "winter/wn-backend-module": "dev-wip/1.2",
+        "winter/wn-cms-module": "dev-wip/1.2",
+        "laravel/framework": "^9.0",
         "wikimedia/composer-merge-plugin": "~2.0.1",
         "twbs/bootstrap": "5.1.3",
         "twig/intl-extra": "^3.3",
@@ -41,10 +41,10 @@
         "winter/wn-docs-plugin": "dev-main"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.4|^9.3.3",
-        "mockery/mockery": "~1.3.3|^1.4.2",
-        "fakerphp/faker": "~1.9",
-        "squizlabs/php_codesniffer": "3.*",
+        "phpunit/phpunit": "^9.5.8",
+        "mockery/mockery": "^1.4.4",
+        "fakerphp/faker": "^1.9.2",
+        "squizlabs/php_codesniffer": "^3.2",
         "php-parallel-lint/php-parallel-lint": "^1.0",
         "dms/phpunit-arraysubset-asserts": "^0.1.0|^0.2.1"
     },
