@@ -23,18 +23,18 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
-        "composer/installers": "~1.11",
-        "laravel/framework": "~6.0",
+        "php": "^8.0.2",
+        "composer/installers": "~1.11.0",
+        "laravel/framework": "^9.0",
         "composer/semver": "^3.2"
+    },
+    "replace": {
+        "october/system": "1.1.*"
     },
     "autoload": {
         "psr-4": {
             "System\\": ""
         }
-    },
-    "extra": {
-        "installer-name": "system"
     },
     "minimum-stability": "dev"
 }
