@@ -317,9 +317,9 @@ return [
         'add_name' => ":name hinzufügen",
         'create' => "Erstellen",
         'create_name' => "Erstelle :name",
-        'update' => "Update",
-        'update_name' => "Update :name",
-        'preview' => 'Vorschaue',
+        'update' => "Speichern",
+        'update_name' => ":name bearbeiten",
+        'preview' => 'Vorschau',
         'preview_name' => 'Vorschau :name',
         'remove' => "Entfernen",
         'remove_name' => ":name entfernen",
