@@ -80,7 +80,7 @@ trait UploadableWidget
             // Use the configured upload path unless it's null, in which case use the user-provided path
             $path = !empty($this->uploadPath) ? $this->uploadPath : Request::input('path');
             $path = MediaLibrary::validatePath($path);
-            $filePath = $path . '/' . $fileName;
+            $filePath = rtrim($path, '/') . '/' . $fileName;
 
             /*
              * getRealPath() can be empty for some environments (IIS)
