@@ -5,6 +5,7 @@ use Url;
 use Lang;
 use File;
 use Event;
+use Config;
 use Backend;
 use BackendMenu;
 use BackendAuth;
@@ -17,7 +18,12 @@ use Cms\Classes\ComponentManager;
 use System\Classes\CombineAssets;
 use Cms\Classes\Theme as CmsTheme;
 use Backend\Classes\WidgetManager;
+use System\Classes\MarkupManager;
 use System\Classes\SettingsManager;
+use Twig\Cache\FilesystemCache as TwigCacheFilesystem;
+use Cms\Twig\Loader as CmsTwigLoader;
+use Cms\Twig\DebugExtension;
+use Cms\Twig\Extension as CmsTwigExtension;
 
 use Winter\Storm\Support\ModuleServiceProvider;
 
@@ -30,8 +36,8 @@ class ServiceProvider extends ModuleServiceProvider
      */
     public function register()
     {
-        parent::register('cms');
-
+        $this->registerConsole();
+        $this->registerTwigParser();
         $this->registerAssetBundles();
         $this->registerComponents();
         $this->registerThemeLogging();
@@ -65,6 +71,58 @@ class ServiceProvider extends ModuleServiceProvider
         if (App::runningInBackend()) {
             $this->bootBackendLocalization();
         }
+    }
+
+    /**
+     * Register command line specifics
+     */
+    protected function registerConsole()
+    {
+        $this->registerConsoleCommand('create.component', \Cms\Console\CreateComponent::class);
+        $this->registerConsoleCommand('create.theme', \Cms\Console\CreateTheme::class);
+
+        $this->registerConsoleCommand('theme.install', \Cms\Console\ThemeInstall::class);
+        $this->registerConsoleCommand('theme.remove', \Cms\Console\ThemeRemove::class);
+        $this->registerConsoleCommand('theme.list', \Cms\Console\ThemeList::class);
+        $this->registerConsoleCommand('theme.use', \Cms\Console\ThemeUse::class);
+        $this->registerConsoleCommand('theme.sync', \Cms\Console\ThemeSync::class);
+    }
+
+    /*
+     * Register Twig Environments and other Twig modifications provided by the module
+     */
+    protected function registerTwigParser()
+    {
+        // Register CMS Twig environment
+        App::singleton('twig.environment.cms', function ($app) {
+            // Load Twig options
+            $useCache = !Config::get('cms.twigNoCache');
+            $isDebugMode = Config::get('app.debug', false);
+            $strictVariables = Config::get('cms.enableTwigStrictVariables', false);
+            $strictVariables = $strictVariables ?? $isDebugMode;
+            $forceBytecode = Config::get('cms.forceBytecodeInvalidation', false);
+
+            $options = [
+                'auto_reload' => true,
+                'debug' => $isDebugMode,
+                'strict_variables' => $strictVariables,
+            ];
+
+            if ($useCache) {
+                $options['cache'] = new TwigCacheFilesystem(
+                    storage_path().'/cms/twig',
+                    $forceBytecode ? TwigCacheFilesystem::FORCE_BYTECODE_INVALIDATION : 0
+                );
+            }
+
+            $twig = MarkupManager::makeBaseTwigEnvironment(new CmsTwigLoader, $options);
+            $twig->addExtension(new CmsTwigExtension);
+            if ($isDebugMode) {
+                $twig->addExtension(new DebugExtension);
+            }
+
+            return $twig;
+        });
     }
 
     /**
