@@ -25,7 +25,7 @@ class PageNode extends TwigNode
     {
         $compiler
             ->addDebugInfo($this)
-            ->write("echo \$this->env->getExtension('Cms\Twig\Extension')->pageFunction();\n")
+            ->write("echo \$this->env->getExtension('Cms\Twig\Extension')->pageFunction(\$context);\n")
         ;
     }
 }
