@@ -25,7 +25,7 @@ class StylesNode extends TwigNode
     {
         $compiler
             ->addDebugInfo($this)
-            ->write("echo \$this->env->getExtension('Cms\Twig\Extension')->assetsFunction('css');\n")
+            ->write("echo \$this->env->getExtension('Cms\Twig\Extension')->assetsFunction(\$context, 'css');\n")
             ->write("echo \$this->env->getExtension('Cms\Twig\Extension')->displayBlock('styles');\n")
         ;
     }
