@@ -6,7 +6,7 @@ if ($.oc === undefined) $.oc = $.wn
 if ($.wn.langMessages === undefined) $.wn.langMessages = {}
 $.wn.langMessages['uk'] = $.extend(
     $.wn.langMessages['uk'] || {},
-    {"markdowneditor":{"formatting":"\u0424\u043e\u0440\u043c\u0430\u0442\u0443\u0432\u0430\u043d\u043d\u044f","quote":"\u0426\u0438\u0442\u0430\u0442\u0430","code":"\u041a\u043e\u0434","header1":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 1","header2":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 2","header3":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 3","header4":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 4","header5":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 5","header6":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 6","bold":"\u0416\u0438\u0440\u043d\u0438\u0439 \u0448\u0440\u0438\u0444\u0442","italic":"\u041a\u0443\u0440\u0441\u0438\u0432","unorderedlist":"\u041d\u0435\u043d\u0443\u043c\u0435\u0440\u043e\u0432\u0430\u043d\u043d\u0438\u0439 \u0441\u043f\u0438\u0441\u043e\u043a","orderedlist":"\u041d\u0443\u043c\u0456\u0440\u043e\u0432\u0430\u043d\u043d\u0438\u0439 \u0441\u043f\u0438\u0441\u043e\u043a","video":"\u0412\u0456\u0434\u0435\u043e","image":"\u0417\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u043d\u044f","link":"\u041f\u043e\u0441\u0438\u043b\u0430\u043d\u043d\u044f","horizontalrule":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u0433\u043e\u0440\u0438\u0437\u043e\u043d\u0442\u0430\u043b\u044c\u043d\u0443 \u0440\u0438\u0441\u043a\u0443","fullscreen":"\u041f\u043e\u0432\u043d\u0438\u0439 \u0435\u043a\u0440\u0430\u043d","preview":"\u041f\u043e\u043f\u0435\u0440\u0435\u0434\u043d\u0456\u0439 \u043f\u0435\u0440\u0435\u0433\u043b\u044f\u0434"},"mediamanager":{"insert_link":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u043f\u043e\u0441\u0438\u043b\u0430\u043d\u043d\u044f","insert_image":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u0437\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u043d\u044f","insert_video":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u0432\u0456\u0434\u0435\u043e","insert_audio":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u0430\u0443\u0434\u0456\u043e","invalid_file_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0444\u0430\u0439\u043b \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438 \u043f\u043e\u0441\u0438\u043b\u0430\u043d\u043d\u044f.","invalid_file_single_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u043e\u0434\u0438\u043d \u0444\u0430\u0439\u043b.","invalid_image_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0437\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u043d\u044f \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438.","invalid_video_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0432\u0456\u0434\u0435\u043e \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438.","invalid_audio_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0430\u0443\u0434\u0456\u043e \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438."},"alert":{"confirm_button_text":"\u041e\u043a","cancel_button_text":"\u0421\u043a\u0430\u0441\u0443\u0432\u0430\u0442\u0438","widget_remove_confirm":"\u0412\u0438\u0434\u0430\u043b\u0438\u0442\u0438 \u0446\u0435\u0439 \u0432\u0456\u0434\u0436\u0435\u0442?"},"datepicker":{"previousMonth":"\u041f\u043e\u043f\u0435\u0440\u0435\u0434\u043d\u0456\u0439 \u043c\u0456\u0441\u044f\u0446\u044c","nextMonth":"\u041d\u0430\u0441\u0442\u0443\u043f\u043d\u0438\u0439 \u043c\u0456\u0441\u044f\u0446\u044c","months":["\u0421\u0456\u0447\u0435\u043d\u044c","\u041b\u044e\u0442\u0438\u0439","\u0411\u0435\u0440\u0435\u0437\u0435\u043d\u044c","\u041a\u0432\u0456\u0442\u0435\u043d\u044c","\u0422\u0440\u0430\u0432\u0435\u043d\u044c","\u0427\u0435\u0440\u0432\u0435\u043d\u044c","\u041b\u0438\u043f\u0435\u043d\u044c","\u0421\u0435\u0440\u043f\u0435\u043d\u044c","\u0412\u0435\u0440\u0435\u0441\u0435\u043d\u044c","\u0416\u043e\u0432\u0442\u0435\u043d\u044c","\u041b\u0438\u0441\u0442\u043e\u043f\u0430\u0434","\u0413\u0440\u0443\u0434\u0435\u043d\u044c"],"weekdays":["\u041d\u0435\u0434\u0456\u043b\u044f","\u041f\u043e\u043d\u0435\u0434\u0456\u043b\u043e\u043a","\u0412\u0456\u0432\u0442\u043e\u0440\u043e\u043a","\u0421\u0435\u0440\u0435\u0434\u0430","\u0427\u0435\u0442\u0432\u0435\u0440","\u041f'\u044f\u0442\u043d\u0438\u0446\u044f","C\u0443\u0431\u043e\u0442\u0430"],"weekdaysShort":["\u041d\u0434","\u041f\u043d","\u0412\u0442","\u0421\u0440","\u0427\u0442","\u041f\u0442","\u0421\u0431"]},"colorpicker":{"last_color":"Use previously selected color","aria_palette":"Color selection area","aria_hue":"Hue selection slider","aria_opacity":"Opacity selection slider"},"filter":{"group":{"all":"\u0432\u0441\u0456"},"scopes":{"apply_button_text":"Apply","clear_button_text":"Clear"},"dates":{"all":"\u0432\u0441\u0456","filter_button_text":"\u0424\u0456\u043b\u044c\u0442\u0440","reset_button_text":"\u0421\u043a\u0438\u043d\u0443\u0442\u0438","date_placeholder":"\u0414\u0430\u0442\u0430","after_placeholder":"\u041f\u0456\u0441\u043b\u044f","before_placeholder":"\u0414\u043e"},"numbers":{"all":"all","filter_button_text":"Filter","reset_button_text":"Reset","min_placeholder":"Min","max_placeholder":"Max"}},"eventlog":{"show_stacktrace":"\u041f\u043e\u043a\u0430\u0437\u0430\u0442\u0438 \u0442\u0440\u0430\u0441\u0443\u0432\u0430\u043d\u043d\u044f \u0441\u0442\u0435\u043a\u0430","hide_stacktrace":"\u041f\u0440\u0438\u0445\u043e\u0432\u0430\u0442\u0438 \u0442\u0440\u0430\u0441\u0443\u0432\u0430\u043d\u043d\u044f \u0441\u0442\u0435\u043a\u0430","tabs":{"formatted":"\u0424\u043e\u0440\u043c\u0430\u0442\u043e\u0432\u0430\u043d\u0438\u0439","raw":"\u041f\u043e\u0447\u0430\u0442\u043a\u043e\u0432\u0438\u0439"},"editor":{"title":"\u0420\u0435\u0434\u0430\u043a\u0442\u043e\u0440 \u0432\u0438\u0445\u0456\u0434\u043d\u043e\u0433\u043e \u043a\u043e\u0434\u0443","description":"\u0412\u0430\u0448\u0430 \u043e\u043f\u0435\u0440\u0430\u0446\u0456\u0439\u043d\u0430 \u0441\u0438\u0441\u0442\u0435\u043c\u0430 \u043f\u043e\u0432\u0438\u043d\u043d\u0430 \u0431\u0443\u0442\u0438 \u043d\u0430\u043b\u0430\u0448\u0442\u043e\u0432\u0430\u043d\u0430 \u043d\u0430 \u043f\u0440\u043e\u0441\u043b\u0443\u0445\u043e\u0432\u0443\u0432\u0430\u043d\u043d\u044f \u0434\u043e \u043e\u0434\u043d\u0456\u0454\u0457 \u0437 \u0446\u0438\u0445 \u0441\u0445\u0435\u043c URL.","openWith":"\u0412\u0456\u0434\u043a\u0440\u0438\u0442\u0438 \u0437\u0430 \u0434\u043e\u043f\u043e\u043c\u043e\u0433\u043e\u044e","remember_choice":"\u0417\u0430\u043f\u0430\u043c'\u044f\u0442\u0430\u0442\u0438 \u043e\u0431\u0440\u0430\u043d\u0438\u0439 \u0432\u0430\u0440\u0456\u0430\u043d\u0442 \u0434\u043b\u044f \u0446\u0456\u0454\u0457 \u0441\u0435\u0441\u0456\u0457","open":"\u0412\u0456\u0434\u043a\u0440\u0438\u0442\u0438","cancel":"\u0421\u043a\u0430\u0441\u0443\u0432\u0430\u0442\u0438"}}}
+    {"markdowneditor":{"formatting":"\u0424\u043e\u0440\u043c\u0430\u0442\u0443\u0432\u0430\u043d\u043d\u044f","quote":"\u0426\u0438\u0442\u0430\u0442\u0430","code":"\u041a\u043e\u0434","header1":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 1","header2":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 2","header3":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 3","header4":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 4","header5":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 5","header6":"\u0417\u0430\u0433\u043e\u043b\u043e\u0432\u043e\u043a 6","bold":"\u0416\u0438\u0440\u043d\u0438\u0439 \u0448\u0440\u0438\u0444\u0442","italic":"\u041a\u0443\u0440\u0441\u0438\u0432","unorderedlist":"\u041d\u0435\u043d\u0443\u043c\u0435\u0440\u043e\u0432\u0430\u043d\u043d\u0438\u0439 \u0441\u043f\u0438\u0441\u043e\u043a","orderedlist":"\u041d\u0443\u043c\u0456\u0440\u043e\u0432\u0430\u043d\u043d\u0438\u0439 \u0441\u043f\u0438\u0441\u043e\u043a","video":"\u0412\u0456\u0434\u0435\u043e","image":"\u0417\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u043d\u044f","link":"\u041f\u043e\u0441\u0438\u043b\u0430\u043d\u043d\u044f","horizontalrule":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u0433\u043e\u0440\u0438\u0437\u043e\u043d\u0442\u0430\u043b\u044c\u043d\u0443 \u0440\u0438\u0441\u043a\u0443","fullscreen":"\u041f\u043e\u0432\u043d\u0438\u0439 \u0435\u043a\u0440\u0430\u043d","preview":"\u041f\u043e\u043f\u0435\u0440\u0435\u0434\u043d\u0456\u0439 \u043f\u0435\u0440\u0435\u0433\u043b\u044f\u0434"},"mediamanager":{"insert_link":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u043f\u043e\u0441\u0438\u043b\u0430\u043d\u043d\u044f","insert_image":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u0437\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u043d\u044f","insert_video":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u0432\u0456\u0434\u0435\u043e","insert_audio":"\u0412\u0441\u0442\u0430\u0432\u0438\u0442\u0438 \u043c\u0435\u0434\u0456\u0430-\u0430\u0443\u0434\u0456\u043e","invalid_file_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0444\u0430\u0439\u043b \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438 \u043f\u043e\u0441\u0438\u043b\u0430\u043d\u043d\u044f.","invalid_file_single_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u043e\u0434\u0438\u043d \u0444\u0430\u0439\u043b.","invalid_image_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0437\u043e\u0431\u0440\u0430\u0436\u0435\u043d\u043d\u044f \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438.","invalid_video_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0432\u0456\u0434\u0435\u043e \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438.","invalid_audio_empty_insert":"\u0411\u0443\u0434\u044c \u043b\u0430\u0441\u043a\u0430, \u043e\u0431\u0435\u0440\u0456\u0442\u044c \u0430\u0443\u0434\u0456\u043e \u0434\u043b\u044f \u0432\u0441\u0442\u0430\u0432\u043a\u0438."},"alert":{"confirm_button_text":"\u041e\u043a","cancel_button_text":"\u0421\u043a\u0430\u0441\u0443\u0432\u0430\u0442\u0438","widget_remove_confirm":"\u0412\u0438\u0434\u0430\u043b\u0438\u0442\u0438 \u0446\u0435\u0439 \u0432\u0456\u0434\u0436\u0435\u0442?"},"datepicker":{"previousMonth":"\u041f\u043e\u043f\u0435\u0440\u0435\u0434\u043d\u0456\u0439 \u043c\u0456\u0441\u044f\u0446\u044c","nextMonth":"\u041d\u0430\u0441\u0442\u0443\u043f\u043d\u0438\u0439 \u043c\u0456\u0441\u044f\u0446\u044c","months":["\u0421\u0456\u0447\u0435\u043d\u044c","\u041b\u044e\u0442\u0438\u0439","\u0411\u0435\u0440\u0435\u0437\u0435\u043d\u044c","\u041a\u0432\u0456\u0442\u0435\u043d\u044c","\u0422\u0440\u0430\u0432\u0435\u043d\u044c","\u0427\u0435\u0440\u0432\u0435\u043d\u044c","\u041b\u0438\u043f\u0435\u043d\u044c","\u0421\u0435\u0440\u043f\u0435\u043d\u044c","\u0412\u0435\u0440\u0435\u0441\u0435\u043d\u044c","\u0416\u043e\u0432\u0442\u0435\u043d\u044c","\u041b\u0438\u0441\u0442\u043e\u043f\u0430\u0434","\u0413\u0440\u0443\u0434\u0435\u043d\u044c"],"weekdays":["\u041d\u0435\u0434\u0456\u043b\u044f","\u041f\u043e\u043d\u0435\u0434\u0456\u043b\u043e\u043a","\u0412\u0456\u0432\u0442\u043e\u0440\u043e\u043a","\u0421\u0435\u0440\u0435\u0434\u0430","\u0427\u0435\u0442\u0432\u0435\u0440","\u041f'\u044f\u0442\u043d\u0438\u0446\u044f","C\u0443\u0431\u043e\u0442\u0430"],"weekdaysShort":["\u041d\u0434","\u041f\u043d","\u0412\u0442","\u0421\u0440","\u0427\u0442","\u041f\u0442","\u0421\u0431"]},"colorpicker":{"last_color":"\u0412\u0438\u043a\u043e\u0440\u0438\u0441\u0442\u043e\u0432\u0443\u0432\u0430\u0442\u0438 \u0440\u0430\u043d\u0456\u0448\u0435 \u0432\u0438\u0431\u0440\u0430\u043d\u0438\u0439 \u043a\u043e\u043b\u0456\u0440","aria_palette":"\u041e\u0431\u043b\u0430\u0441\u0442\u044c \u0432\u0438\u0431\u043e\u0440\u0443 \u043a\u043e\u043b\u044c\u043e\u0440\u0443","aria_hue":"\u041f\u043e\u0432\u0437\u0443\u043d\u043e\u043a \u0432\u0438\u0431\u043e\u0440\u0443 \u0432\u0456\u0434\u0442\u0456\u043d\u043a\u0443","aria_opacity":"\u041f\u043e\u0432\u0437\u0443\u043d\u043e\u043a \u0432\u0438\u0431\u043e\u0440\u0443 \u043f\u0440\u043e\u0437\u043e\u0440\u043e\u0441\u0442\u0456"},"filter":{"group":{"all":"\u0432\u0441\u0456"},"scopes":{"apply_button_text":"\u0417\u0430\u0441\u0442\u043e\u0441\u0443\u0432\u0430\u0442\u0438","clear_button_text":"\u041e\u0447\u0438\u0441\u0442\u0438\u0442\u0438"},"dates":{"all":"\u0432\u0441\u0456","filter_button_text":"\u0424\u0456\u043b\u044c\u0442\u0440","reset_button_text":"\u0421\u043a\u0438\u043d\u0443\u0442\u0438","date_placeholder":"\u0414\u0430\u0442\u0430","after_placeholder":"\u041f\u0456\u0441\u043b\u044f","before_placeholder":"\u0414\u043e"},"numbers":{"all":"\u0432\u0441\u0456","filter_button_text":"\u0424\u0456\u043b\u044c\u0442\u0440","reset_button_text":"\u0421\u043a\u0438\u0434\u0430\u043d\u043d\u044f","min_placeholder":"Min","max_placeholder":"Max"}},"eventlog":{"show_stacktrace":"\u041f\u043e\u043a\u0430\u0437\u0430\u0442\u0438 \u0442\u0440\u0430\u0441\u0443\u0432\u0430\u043d\u043d\u044f \u0441\u0442\u0435\u043a\u0430","hide_stacktrace":"\u041f\u0440\u0438\u0445\u043e\u0432\u0430\u0442\u0438 \u0442\u0440\u0430\u0441\u0443\u0432\u0430\u043d\u043d\u044f \u0441\u0442\u0435\u043a\u0430","tabs":{"formatted":"\u0424\u043e\u0440\u043c\u0430\u0442\u043e\u0432\u0430\u043d\u0438\u0439","raw":"\u041f\u043e\u0447\u0430\u0442\u043a\u043e\u0432\u0438\u0439"},"editor":{"title":"\u0420\u0435\u0434\u0430\u043a\u0442\u043e\u0440 \u0432\u0438\u0445\u0456\u0434\u043d\u043e\u0433\u043e \u043a\u043e\u0434\u0443","description":"\u0412\u0430\u0448\u0430 \u043e\u043f\u0435\u0440\u0430\u0446\u0456\u0439\u043d\u0430 \u0441\u0438\u0441\u0442\u0435\u043c\u0430 \u043f\u043e\u0432\u0438\u043d\u043d\u0430 \u0431\u0443\u0442\u0438 \u043d\u0430\u043b\u0430\u0448\u0442\u043e\u0432\u0430\u043d\u0430 \u043d\u0430 \u043f\u0440\u043e\u0441\u043b\u0443\u0445\u043e\u0432\u0443\u0432\u0430\u043d\u043d\u044f \u0434\u043e \u043e\u0434\u043d\u0456\u0454\u0457 \u0437 \u0446\u0438\u0445 \u0441\u0445\u0435\u043c URL.","openWith":"\u0412\u0456\u0434\u043a\u0440\u0438\u0442\u0438 \u0437\u0430 \u0434\u043e\u043f\u043e\u043c\u043e\u0433\u043e\u044e","remember_choice":"\u0417\u0430\u043f\u0430\u043c'\u044f\u0442\u0430\u0442\u0438 \u043e\u0431\u0440\u0430\u043d\u0438\u0439 \u0432\u0430\u0440\u0456\u0430\u043d\u0442 \u0434\u043b\u044f \u0446\u0456\u0454\u0457 \u0441\u0435\u0441\u0456\u0457","open":"\u0412\u0456\u0434\u043a\u0440\u0438\u0442\u0438","cancel":"\u0421\u043a\u0430\u0441\u0443\u0432\u0430\u0442\u0438"}}}
 );
 
 //! moment.js locale configuration v2.22.2
