@@ -1,12 +1,11 @@
 <?php namespace System\Classes;
 
-use Twig;
+use App;
 use Markdown;
 use System\Models\MailPartial;
 use System\Models\MailTemplate;
 use System\Models\MailBrandSetting;
 use System\Helpers\View as ViewHelper;
-use System\Twig\MailPartialTokenParser;
 use TijsVerkoyen\CssToInlineStyles\CssToInlineStyles;
 
 /**
@@ -48,11 +47,6 @@ class MailManager
      * @var bool Internal marker for rendering mode
      */
     protected $isHtmlRenderMode = false;
-
-    /**
-     * @var bool Internal marker for booting custom twig extensions.
-     */
-    protected $isTwigStarted = false;
 
     /**
      * Same as `addContentToMailer` except with raw content.
@@ -115,11 +109,6 @@ class MailManager
     protected function addContentToMailerInternal($message, $template, $data, $plainOnly = false)
     {
         /*
-         * Start twig transaction
-         */
-        $this->startTwig();
-
-        /*
          * Inject global view variables
          */
         $globalVars = ViewHelper::getGlobalVars();
@@ -130,14 +119,14 @@ class MailManager
         /*
          * Subject
          */
-        $swiftMessage = $message->getSwiftMessage();
+        $symfonyMessage = $message->getSymfonyMessage();
 
-        if (empty($swiftMessage->getSubject())) {
-            $message->subject(Twig::parse($template->subject, $data));
+        if (empty($symfonyMessage->getSubject())) {
+            $message->subject($this->renderTwig($template->subject, $data));
         }
 
         $data += [
-            'subject' => $swiftMessage->getSubject()
+            'subject' => $symfonyMessage->getSubject()
         ];
 
         if (!$plainOnly) {
@@ -146,7 +135,7 @@ class MailManager
              */
             $html = $this->renderTemplate($template, $data);
 
-            $message->setBody($html, 'text/html');
+            $message->html($html);
         }
 
         /*
@@ -154,12 +143,7 @@ class MailManager
          */
         $text = $this->renderTextTemplate($template, $data);
 
-        $message->addPart($text, 'text/plain');
-
-        /*
-         * End twig transaction
-         */
-        $this->stopTwig();
+        $message->text($text);
     }
 
     //
@@ -202,7 +186,7 @@ class MailManager
             $html = $this->renderTwig($template->layout->content_html, [
                 'content' => $html,
                 'css' => $template->layout->content_css,
-                'brandCss' => $css
+                'brandCss' => $css,
             ] + (array) $data);
 
             $css .= PHP_EOL . $template->layout->content_css;
@@ -276,56 +260,13 @@ class MailManager
     }
 
     /**
-     * Internal helper for rendering Twig
+     * Internal helper for rendering Twig using the mailer Twig environment
      */
-    protected function renderTwig($content, $data = [])
+    protected function renderTwig(string $content, array $data = []): string
     {
-        if ($this->isTwigStarted) {
-            return Twig::parse($content, $data);
-        }
-
-        $this->startTwig();
-
-        $result = Twig::parse($content, $data);
-
-        $this->stopTwig();
-
-        return $result;
-    }
-
-    /**
-     * Temporarily registers mail based token parsers with Twig.
-     * @return void
-     */
-    protected function startTwig()
-    {
-        if ($this->isTwigStarted) {
-            return;
-        }
-
-        $this->isTwigStarted = true;
-
-        $markupManager = MarkupManager::instance();
-        $markupManager->beginTransaction();
-        $markupManager->registerTokenParsers([
-            new MailPartialTokenParser
-        ]);
-    }
-
-    /**
-     * Indicates that we are finished with Twig.
-     * @return void
-     */
-    protected function stopTwig()
-    {
-        if (!$this->isTwigStarted) {
-            return;
-        }
-
-        $markupManager = MarkupManager::instance();
-        $markupManager->endTransaction();
-
-        $this->isTwigStarted = false;
+        return App::make('twig.environment.mailer')
+                ->createTemplate($content)
+                ->render($data);
     }
 
     //
