@@ -48,7 +48,7 @@ class WinterManifest extends \Illuminate\Console\Command
     /**
      * Execute the console command.
      *
-     * @return mixed
+     * @return void
      */
     public function handle()
     {
