@@ -42,7 +42,7 @@ class WinterVersion extends \Illuminate\Console\Command
     /**
      * Execute the console command.
      *
-     * @return mixed
+     * @return void
      */
     public function handle()
     {
@@ -65,7 +65,7 @@ class WinterVersion extends \Illuminate\Console\Command
             $this->info('*** Detected Winter CMS build ' . $build['build'] . '.');
         }
 
-        if ($this->option('changes')) {
+        if (!empty($build['changes']) && $this->option('changes')) {
             $this->line('');
             $this->comment('We have detected the following modifications:');
 
