@@ -25,11 +25,9 @@ class Loader implements TwigLoaderInterface
     protected $cache = [];
 
     /**
-     * Gets the path of a view file
-     * @param  string $name
-     * @return string
+     * Gets the path of a view file.
      */
-    protected function findTemplate($name)
+    protected function findTemplate(string $name): string
     {
         $finder = App::make('view')->getFinder();
 
@@ -45,27 +43,43 @@ class Loader implements TwigLoaderInterface
         return $this->cache[$name] = $path;
     }
 
-    public function getSourceContext($name)
+    /**
+     * Returns the Twig content string.
+     * This step is cached internally by Twig.
+     */
+    public function getSourceContext(string $name): TwigSource
     {
         return new TwigSource(File::get($this->findTemplate($name)), $name);
     }
 
-    public function getCacheKey($name)
+    /**
+     * Returns the Twig cache key.
+     */
+    public function getCacheKey(string $name): string
     {
         return $this->findTemplate($name);
     }
 
-    public function isFresh($name, $time)
+    /**
+     * Determines if the content is fresh.
+     */
+    public function isFresh(string $name, int $time): bool
     {
         return File::lastModified($this->findTemplate($name)) <= $time;
     }
 
-    public function getFilename($name)
+    /**
+     * Returns the file name of the loaded template.
+     */
+    public function getFilename(string $name): string
     {
         return $this->findTemplate($name);
     }
 
-    public function exists($name)
+    /**
+     * Checks that the template exists.
+     */
+    public function exists(string $name): bool
     {
         try {
             $this->findTemplate($name);
