@@ -1,11 +1,8 @@
 <?php
 
-/*
- * (c) Boosteur.com - 2022
- */
-
 return [
-  /*
+
+    /*
      |--------------------------------------------------------------------------
      | Debugbar Status
      |--------------------------------------------------------------------------
@@ -15,9 +12,9 @@ return [
      |
      */
 
-  'enabled' => env('DEBUGBAR_ENABLED', null),
+    'enabled' => env('DEBUGBAR_ENABLED', null),
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Allow public access
      |--------------------------------------------------------------------------
@@ -33,9 +30,9 @@ return [
      |
      */
 
-  'allow_public_access' => true,
+    'allow_public_access' => false,
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Store all requests
      |--------------------------------------------------------------------------
@@ -54,9 +51,10 @@ return [
      |
      */
 
-  'store_all_requests' => false,
+    'store_all_requests' => false,
 
-  /*
+
+    /*
      |--------------------------------------------------------------------------
      | Routes to exclude
      |--------------------------------------------------------------------------
@@ -65,11 +63,11 @@ return [
      |
      */
 
-  'except' => [
-    'telescope*',
-  ],
+    'except' => [
+        'telescope*'
+    ],
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Storage settings
      |--------------------------------------------------------------------------
@@ -82,15 +80,15 @@ return [
      |
      */
 
-  'storage' => [
-    'enabled'    => true,
-    'driver'     => 'file', // redis, file, pdo, custom
-    'path'       => storage_path('debugbar'), // For file driver
-    'connection' => null,   // Leave null for default connection (Redis/PDO)
-    'provider'   => '', // Instance of StorageInterface for custom driver
-  ],
+    'storage' => [
+        'enabled'    => true,
+        'driver'     => 'file', // redis, file, pdo, custom
+        'path'       => storage_path('debugbar'), // For file driver
+        'connection' => null,   // Leave null for default connection (Redis/PDO)
+        'provider'   => '' // Instance of StorageInterface for custom driver
+    ],
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Vendors
      |--------------------------------------------------------------------------
@@ -104,9 +102,9 @@ return [
      |
      */
 
-  'include_vendors' => true,
+    'include_vendors' => true,
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Capture Ajax Requests
      |--------------------------------------------------------------------------
@@ -117,10 +115,10 @@ return [
      | Optionally, you can also send ServerTiming headers on ajax requests for the Chrome DevTools.
      */
 
-  'capture_ajax'    => true,
-  'add_ajax_timing' => false,
+    'capture_ajax' => true,
+    'add_ajax_timing' => false,
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Custom Error Handler for Deprecated warnings
      |--------------------------------------------------------------------------
@@ -130,9 +128,9 @@ return [
      |
      */
 
-  'error_handler' => false,
+    'error_handler' => false,
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Clockwork integration
      |--------------------------------------------------------------------------
@@ -142,9 +140,9 @@ return [
      |
      */
 
-  'clockwork' => false,
+    'clockwork' => false,
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | DataCollectors
      |--------------------------------------------------------------------------
@@ -153,32 +151,32 @@ return [
      |
      */
 
-  'collectors' => [
-    'phpinfo'         => true,  // Php version
-    'messages'        => true,  // Messages
-    'time'            => true,  // Time Datalogger
-    'memory'          => true,  // Memory usage
-    'exceptions'      => true,  // Exception displayer
-    'log'             => true,  // Logs from Monolog (merged in messages if enabled)
-    'db'              => true,  // Show database (PDO) queries and bindings
-    'views'           => true,  // Views with their data
-    'route'           => true,  // Current route information
-    'auth'            => false, // Display Laravel authentication status
-    'gate'            => true, // Display Laravel Gate checks
-    'session'         => true,  // Display session data
-    'symfony_request' => true,  // Only one can be enabled..
-    'mail'            => true,  // Catch mail messages
-    'laravel'         => false, // Laravel version and environment
-    'events'          => true, // All events fired
-    'default_request' => false, // Regular or special Symfony request logger
-    'logs'            => false, // Add the latest log messages
-    'files'           => false, // Show the included files
-    'config'          => false, // Display config settings
-    'cache'           => false, // Display cache events
-    'models'          => true, // Display models
-  ],
+    'collectors' => [
+        'phpinfo'         => true,  // Php version
+        'messages'        => true,  // Messages
+        'time'            => true,  // Time Datalogger
+        'memory'          => true,  // Memory usage
+        'exceptions'      => true,  // Exception displayer
+        'log'             => true,  // Logs from Monolog (merged in messages if enabled)
+        'db'              => true,  // Show database (PDO) queries and bindings
+        'views'           => true,  // Views with their data
+        'route'           => true,  // Current route information
+        'auth'            => false, // Display Laravel authentication status
+        'gate'            => true, // Display Laravel Gate checks
+        'session'         => true,  // Display session data
+        'symfony_request' => true,  // Only one can be enabled..
+        'mail'            => true,  // Catch mail messages
+        'laravel'         => false, // Laravel version and environment
+        'events'          => true, // All events fired
+        'default_request' => false, // Regular or special Symfony request logger
+        'logs'            => false, // Add the latest log messages
+        'files'           => false, // Show the included files
+        'config'          => false, // Display config settings
+        'cache'           => false, // Display cache events
+        'models'          => true, // Display models
+    ],
 
-  /*
+    /*
      |--------------------------------------------------------------------------
      | Extra options
      |--------------------------------------------------------------------------
@@ -187,34 +185,34 @@ return [
      |
      */
 
-  'options' => [
-    'auth' => [
-      'show_name' => true,   // Also show the users name/email in the debugbar
-    ],
-    'db' => [
-      'with_params' => true,   // Render SQL with the parameters substituted
-      'backtrace'   => true,   // Use a backtrace to find the origin of the query in your files.
-      'timeline'    => false,  // Add the queries to the timeline
-      'explain'     => [                 // Show EXPLAIN output on queries
-        'enabled' => false,
-        'types'   => ['SELECT'],     // ['SELECT', 'INSERT', 'UPDATE', 'DELETE']; for MySQL 5.6.3+
-      ],
-      'hints' => true,    // Show hints for common mistakes
+    'options' => [
+        'auth' => [
+            'show_name' => true,   // Also show the users name/email in the debugbar
+        ],
+        'db' => [
+            'with_params'       => true,   // Render SQL with the parameters substituted
+            'backtrace'         => true,   // Use a backtrace to find the origin of the query in your files.
+            'timeline'          => false,  // Add the queries to the timeline
+            'explain' => [                 // Show EXPLAIN output on queries
+                'enabled' => false,
+                'types' => ['SELECT'],     // ['SELECT', 'INSERT', 'UPDATE', 'DELETE']; for MySQL 5.6.3+
+            ],
+            'hints'             => true,    // Show hints for common mistakes
         ],
         'mail' => [
-          'full_log' => false,
+            'full_log' => false
         ],
         'views' => [
-          'data' => false,    //Note: Can slow down the application, because the data can be quite large..
+            'data' => false,    //Note: Can slow down the application, because the data can be quite large..
         ],
         'route' => [
-          'label' => true,  // show complete route on bar
+            'label' => true  // show complete route on bar
         ],
         'logs' => [
-          'file' => null,
+            'file' => null
         ],
         'cache' => [
-          'values' => true, // collect cache values
+            'values' => true // collect cache values
         ],
     ],
 
